change how we are considering the alarm to call scheduler again

+++ proc.h
@@ -52,7 +52,7 @@ struct proc {
   int priority;                // Process priority (scheduling)
   int maxTimerTicks;
   int currentTimerTicks;
-  int execTime;
+  float execTime;
 };
 
 // Process memory is laid out contiguously, low addresses first:
